sprint1/fast_api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0

psycopg2-binary==2.9.6
psycopg2-pool==1.1

pydantic==1.10.7

pyyaml==6.0

pytest==7.3.1
httpx==0.27.0
pytest-asyncio==0.23.6

gunicorn==20.1.0